Move r2rml implementations
