+++ src/main/java/r2graph/r2rml/R2RMLMap.java
@@ -1,4 +1,4 @@
-package r2graph.configmap.impl;
+package r2graph.r2rml;
 
 import r2graph.configmap.ConfigMap;
 import r2graph.configmap.EntityMap;
