+++ src/main/java/r2graph/r2rml/TriplesMap.java
@@ -1,4 +1,4 @@
-package r2graph.configmap.impl;
+package r2graph.r2rml;
 
 import r2graph.configmap.EntityMap;
 import r2graph.configmap.PredicateObjectMap;
