+++ src/test/java/r2graph/r2rml/R2RMLMapTest.java
@@ -1,7 +1,8 @@
-package r2graph.configmap.impl;
+package r2graph.r2rml;
 
 import r2graph.configmap.AbstractConfigMapTest;
 import r2graph.configmap.ConfigMap;
+import r2graph.r2rml.R2RMLMap;
 
 public class R2RMLMapTest extends AbstractConfigMapTest {
     @Override
