+++ src/test/java/r2graph/r2rml/TriplesMapTest.java
@@ -1,7 +1,8 @@
-package r2graph.configmap.impl;
+package r2graph.r2rml;
 
 import r2graph.configmap.AbstractEntityMapTest;
 import r2graph.configmap.EntityMap;
+import r2graph.r2rml.TriplesMap;
 
 public class TriplesMapTest extends AbstractEntityMapTest {
     @Override
